fix(perf): 메인 직접 매출 ₩0 버그 + 캠페인 편집 페이지 폐지
1) 히어로 '메인 상품 직접 매출 ₩0'·'미지정'·증분Top10 전부 회색 버그:
   promotion_measurement 가 빈 promotion_main_products 테이블로 is_main 판정 → 항상 false.
   plan_vs_actual 과 동일하게 campaign_plans.main_product_ids(null=전체)+정규화 이름 매칭으로
   통일(0056). 검증(벌크UP): 메인 직접 매출 ₩0→1.41억, 메인 측정행 0→10개.
   '지정하기' 링크도 편집페이지 → 플랜으로(메인은 플랜에서 옵션으로 지정).
2) 캠페인 정보 편집 페이지 폐지: 메인은 플랜, 실공헌/광고는 성과화면, 병합 제거됨 →
   편집 페이지가 사실상 불필요. 캠페인 삭제만 상세 헤더로 이동(DeleteCampaignButton).
   업로드 후 이동도 상세로. (이름·기간·목적은 캠페인 작성 단계에서 설정)

tsc·build·테스트 통과. 0056 prod 적용·실데이터 검증.
